api/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.1
cryptography==41.0.5
daphne==4.0.0
distro==1.8.0
Django==4.2.1
djangorestframework==3.14.0
drf-yasg==1.20.0
exceptiongroup==1.2.0
execnet==2.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.0.1
openai==1.1.1
packaging==23.2
pluggy==1.3.0
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
PyJWT==2.8.0
pyOpenSSL==23.3.0
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-forked==1.4.0
pytest-mock==3.10.0
pytest-xdist==2.5.0
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
zope.interface==6.1
